requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1637088766493/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1632748415174/work
akro==0.0.8
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach==5.0.0
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1640686991047/work
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1649184113124/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
circle-fit==0.1.3
click @ file:///home/conda/feedstock_root/build_artifacts/click_1648793466560/work
cloudpickle==2.0.0
cma==2.7.0
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1634254711879/work
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.0
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Deprecated==1.2.13
dm-tree==0.1.6
dowel==0.0.3
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649908585034/work
fasteners==0.17.3
fastjsonschema==2.15.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.30.0
-e git+ssh://git@github.com/livcms/garage-ParticleTracking.git@7fac247f26035964bb3331f33b68221b514e1e79#egg=garage
gast==0.5.3
glfw==2.5.1
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1650592442422/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1630497468950/work
google-pasta==0.2.0
grpcio==1.43.0
gym==0.23.0
gym-notices==0.0.6
h5py==3.6.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
imageio==2.16.1
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1648728291958/work
importlib-resources==5.4.0
ipykernel==6.13.0
ipython==7.33.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.7.0
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2==3.1.1
joblib==1.1.0
jsonschema @ file:///tmp/build/80754af9/jsonschema_1650008095244/work
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console @ file:///opt/conda/conda-bld/jupyter_console_1647002188872/work
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
Mako==1.2.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1637220118004/work
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
metaworld==0.0.1.dev0
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
msgpack==1.0.3
mujoco-py==2.1.2.14
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
nbclient==0.6.2
nbconvert==6.5.0
nbformat @ file:///tmp/build/80754af9/nbformat_1649844918397/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847911453/work
networkx==2.6.3
notebook==6.4.11
numexpr==2.8.1
numpy==1.21.5
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1643507977997/work
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pygame==2.1.2
pyglet==1.5.22
Pygments==2.12.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyWavelets==1.3.0
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1649078897110/work
QtPy==2.1.0
ray==1.11.0
redis==4.1.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1637781155505/work
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
setproctitle==1.2.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soupsieve==2.3.2.post1
tables==3.7.0
tabulate==0.8.9
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1644015949874/work/tensorboard-2.8.0-py3-none-any.whl
tensorboard-data-server==0.6.1
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorboardX==2.5
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-probability==0.16.0
termcolor==1.1.0
terminado==0.13.3
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
torch==1.7.1
torch-geometric==2.0.4
torchvision==0.8.2
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm==4.64.0
trackml @ file:///home/lhv14/trackml-library
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1650370875435/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1648847766614/work
widgetsnbextension==3.6.0
wrapt==1.14.0
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
